# internal/authorization/repository/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.19.1

package queries

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type DefaultUserNotificationParam struct {
	UserID    int32
	CreatedAt pgtype.Timestamp
	Params    []byte
}

type Task struct {
	ID           int32
	CreatedAt    pgtype.Timestamp
	Message      string
	UserID       int32
	RequiredTime pgtype.Interval
	Periodic     bool
	Done         bool
	Archived     bool
}

type TimetableTask struct {
	ID           int32
	CreatedAt    pgtype.Timestamp
	Text         string
	Description  pgtype.Text
	UserID       int32
	Start        pgtype.Timestamp
	Finish       pgtype.Timestamp
	Done         bool
	TaskID       int32
	Notification []byte
}

type User struct {
	ID           int32
	Email        pgtype.Text
	PasswordHash pgtype.Text
	TgID         pgtype.Int4
}